requirements.txt
quart
pyrofork
python-dotenv
parse-torrent-title
pyjwt
tgcrypto
motor
aiofiles
dnspython
requests